Support a custom working directory
This change allows users to optionally specify
which directory has the git repository.

In rare cases, the directory containing the
desired repository does not exist within
the github workspace on a runner. When this
occurs, this action is no longer useable.

The `working-directory` option within a workflow
is only supported for `run` steps and not `uses`:
https://stackoverflow.com/questions/67299058/running-github-actions-for-uses-in-another-directory